Extract ResultExtension to separate file
