+++ src/extensions/mod.rs
@@ -1,14 +1,2 @@
-pub trait ResultExtensions {
-    fn log_on_error(&self, message: &str);
-}
-
-impl<T, TError> ResultExtensions for Result<T, TError> {
-    fn log_on_error(&self, message: &str) {
-        match self {
-            Err(_error) => {
-                println!("{}", message);
-            },
-            _ => {}
-        }
-    }
-}
+mod result_ext;
+pub use self::result_ext::ResultExtensions;
